-!- gestion tmx | segment fault

+++ game/code/squareEntity.cc
@@ -14,7 +14,7 @@
 
 namespace swiftness
 {
-    Square::Square(gf::Vector2f position, float size, gf::Color4f color, float gravity)
+    Square::Square(gf::Vector2i position, float size, gf::Color4f color, float gravity)
         : m_position(position), m_velocity(0, 0), m_size(size), m_color(color), gravity(GRAVITY_SQUARE), m_jump(false), nb_jumps(0)
     {
     }
