+++ game/code/squareEntity.h
@@ -15,7 +15,7 @@ namespace swiftness
     class Square
     {
     public:
-        Square(gf::Vector2f position, float size, gf::Color4f color, float gravity);
+        Square(gf::Vector2i position, float size, gf::Color4f color, float gravity);
 
         /**
          * @brief Récupère la position du carré
